Move task data to its own file.
